+++ src/main/java/com/mobble/mobbleserver/application/address/port/required/AddressWritePort.java
@@ -1,0 +1,8 @@
+package com.mobble.mobbleserver.application.address.port.required;
+
+import com.mobble.mobbleserver.domain.address.Address;
+
+public interface AddressWritePort {
+
+    Address save(Address address);
+}
