+++ src/main/java/com/mobble/mobbleserver/domain/address/Address.java
@@ -1,8 +1,8 @@
-package com.mobble.mobbleserver.refactor.adress.entity;
+package com.mobble.mobbleserver.domain.address;
 
 import com.mobble.mobbleserver.common.baseEntity.BaseEntity;
 import com.mobble.mobbleserver.domain.club.core.Club;
-import com.mobble.mobbleserver.refactor.adress.dto.request.AddressRequestDto;
+import com.mobble.mobbleserver.infrastructure.web.address.dto.request.AddressRequestDto;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Builder;
