+++ src/main/java/com/mobble/mobbleserver/domain/club/core/Club.java
@@ -4,7 +4,7 @@ import com.mobble.mobbleserver.common.baseEntity.BaseEntity;
 import com.mobble.mobbleserver.domain.chat.room.ClubRoomInfo;
 import com.mobble.mobbleserver.global.exception.common.DomainException;
 import com.mobble.mobbleserver.global.exception.errorCode.club.ClubErrorCode;
-import com.mobble.mobbleserver.refactor.adress.entity.Address;
+import com.mobble.mobbleserver.domain.address.Address;
 import com.mobble.mobbleserver.refactor.clubCategory.entity.ClubCategory;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
